restructured crypto-transactions-history component, added afterAll, removed afterEach
